wlan_utils: Move wifi and caldata scripts to a separate project
For openwrt19.07 kernel5.4 upgrade, wifi and cal-data scripts are
separated from Openwrt base-files and moved to a seperate package.
As this is common project location, this project can be used by
opensource WiFi project also.
The DEPENDS target flag check will ensure the package is enabled
only for kernel5.4 and beyond.

Change-Id: I3384de0605919b5d846a5898cffb892ea1f4d313
Signed-off-by: Naresh Parati <cnparat@codeaurora.org>