wlan utils: qca-cnss: Add md5 value to PKG_BUILD_ID
Adding md5 value for PKG_BUILD_ID.
This is required after openwrt upgrade to 23.05.

This change is using sha1 and md5 for --build-id generation
in stripped and unstripped ipq_cnss2.ko.

Previously this was done by the gcc compiler but after the
upgrade this needs to be added for crashscope parsing.

Change-Id: Ia3896357d3ab188210e11cb61824f0da65629008
Signed-off-by: Raghavendran Madhavan <quic_raghmadh@quicinc.com>